backend/requirements.txt
flask
flask-cors
flask-socketio
requests
pandas
numpy
scikit-learn
eventlet
Pillow
twilio
boto3
google-cloud-storage
ultralytics
torch
torchvision
python-dotenv
Flask-SQLAlchemy  
 